Inventors_Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
requests
python-dotenv
openai
alembic